+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Show creator for created cards on play (#49)
+- HSReplay: Add warning if tag is missing required entity id
 
 ### Fixed
 - Fix fullscreen callback in Launcher not being called